# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: BANCO Asset Marketplace API — Cars, Real Estate, Industrial
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [INVALID_DATA, NOT_FOUND, UNAUTHORIZED, INTERNAL_ERROR, FORBIDDEN, RATE_LIMITED]
        message:
          type: string
    Meta:
      type: object
      properties:
        cursor:
          type: string
        has_next:
          type: boolean
        total:
          type: integer
...
    InvoiceLineItem:
      type: object
      required: [label, amount]
      properties:
        label:
          type: string
        amount:
          type: string
    Invoice:
      type: object
      required:
        [
          id,
          invoice_number,
          amount,
          status,
          transaction_id,
          transaction_type,
          description,
          line_items,
          issued_at,
          created_at,
        ]
      properties:
        id:
          type: string
        invoice_number:
          type: string
        amount:
          type: string
        status:
          type: string
          enum: [paid, void]
        transaction_id:
          type: string
        transaction_type:
          type: string
          nullable: true
          enum:
            [
              wallet_topup,
              boost_charge,
              subscription_charge,
              lead_charge,
              refund,
              adjustment,
            ]
        description:
          type: string
          nullable: true
        line_items:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        issued_at:
          type: string
          nullable: true
        created_at:
          type: string
          nullable: true
    BillingReportLine:
      type: object
      required: [type, total, count]
      properties:
        type:
          type: string
          enum:
            [
              wallet_topup,
              boost_charge,
              subscription_charge,
              lead_charge,
              refund,
              adjustment,
            ]
        total:
          type: string
        count:
          type: integer
    BillingReport:
      type: object
      required:
        [
          month,
          currency,
          total_charged,
          total_topped_up,
          transaction_count,
          by_type,
        ]
      properties:
        month:
          type: string
        currency:
          type: string
          enum: [EGP]
        total_charged:
          type: string
        total_topped_up:
          type: string
        transaction_count:
          type: integer
        by_type:
          type: array
          items:
            $ref: "#/components/schemas/BillingReportLine"
...
paths:
...
  /v1/billing/invoices:
    get:
      operationId: listInvoices
      summary: List the caller's invoices (newest first)
      tags: [billing]
      security:
        - ClerkAuth: []
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: Invoice page
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invoice"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
...
  /v1/billing/invoices/{id}:
    get:
      operationId: getInvoice
      summary: Get a single invoice the caller owns
      tags: [billing]
      security:
        - ClerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Invoice"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
...
  /v1/billing/report:
    get:
      operationId: getBillingReport
      summary: Monthly billing summary computed from the ledger
      tags: [billing]
      security:
        - ClerkAuth: []
      parameters:
        - name: month
          in: query
          required: false
          description: Report month as YYYY-MM (UTC). Defaults to the current month.
          schema:
            type: string
      responses:
        "200":
          description: Billing report
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/BillingReport"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
...